Realloc-without-new-REBSER, fix memory accounting
An earlier commit switched it so that when you expand a series
you do not allocate a new REBSER to do it.  This expands the
ability to resize just the data array inside of a series while reusing
the series header that had previously been allocated.

(Motivation for this was to distinguish in tracking the difference
between a new series being created and an existing series being
resized, but there are several other benefits.)

This also corrects an issue in accounting for the size of memory
allocations.  The FFI code was the first example of creating series
with element sizes that did not divide evenly into 2048.  Because
of this, the logic for knowing how big a series allocation was
would be thrown off.  The actual size of an allocation is not
stored...just the element capacity and the 8-bit "wide" (element
size).  These are multiplied together, and the remainder will
be lost if placed in a 2k rounded up chunk.

Rather than store the precise element size, this sets a bit as
to whether the rounding was done so that the same rounding
calculation can be done at the time of free to get the correct
size.  This means the data used for bounding memory
allocations or triggering the GC will balance to zero when all
the series are freed...and won't be perpetually skewing off.