More aggressively free temp series vs. wait for GC
These temporary series were created which were marked
as "leaks" by the formal managed memory model.  They would
be freed by the garbage collector later.  But freeing them when
they are known not to be used saves reclaims the memory
sooner (and stops the leak alert in the build that has it...)