Overhaul of binding routine names and parameters
In first trying to debug the binding routines, I was quickly confused by
the way the parameters were named and specified.  There would be
a routine with a name like "Bind_BlockXXX" that would take a REBVAL*
that was named block.  You would pass it a REB_BLOCK and it would
crash.  You would then see the code doing things like testing to see
if (IS_WORD(block)) and wonder what was going on.

After a while I realized that the routines were being passed pointers
to the head of a block's data series.  So the 'block' represented the
contents of the block...not a value holding the block series.

The strategy of attack I used to fix this was to give the routines a name
with the word "Array" in it, so "Bind_Array_X" instead of "Bind_Block_X".
That gives a pretty strong hint that even though you are passing a
REBVAL*, it better be the kind of pointer you can increment to get
another REBVAL.  (I considered "Bind_Values_X" but that could be
interpreted just to mean "bind the values in the subtree under the
singular REBVAL* I am passing.)

Secondly, instead of calling the parameter `REBVAL *block` I call it
`REBVAL value[]`.  For a function parameter, C compilers treat that
just like `REBVAL *value`...but it helps make a suggestion in the
prototype that it's an array.  By calling it "value" instead of "values"
it can be reused for iteration and speaking about one element at a
time...vs having a separate variable `value` that could be confused
with `values` and have the wrong one updated/used.

(Note: Where you see &value[0] used instead of passing just value,
that is not a mistake.  That is to indicate at the call site that it is not
a singular value, rather the first element of an array being passed.)

Spreading that decision through led to some changes, and as long
as I was changing things I did a little cleanup of parameter ordering
and some other trivial stuff that I think pushes it closer to being
understandable.  That included turning some helper implementation
functions static and giving them names suggesting they were not
meant to be APIs beyond the private implementation of the function.