Axe COPY_VALUES and COPY_BLK_PART macros
In the landscape of routines you might-consider-as-candidates to call,
these macros are a distraction.  They are wrappers for memcpy (but
with COPY_VALUES the parameter is out of order) and as such are
probably not what callers want.  It's more clear to just inline these
calls at their callsites, which are the places implementing the actual
functionality named.

Includes a small cleanup of Copy_Stack_Values() during the inlining.