Do not preserve data in restarted Widen_String
One call to Widen_Series is called on an in-progress constructed
string where the construction is going to start over, and when it
starts over it uses a memcpy.  So there is no point in preserving
the existing data copied so far (which would just be done with a
memcpy itself) and the string is also not terminated yet.

It would be possible to terminate the series so it was valid, use
the memory preservation inside Widen_String, and then do a
memcpy of just the additional part of the source string.  But it
might be useful for other routines to be able to change the
width of a string without checking or preserving its contents, so
this passes a flag.

This change was called out as needed by a small modification
to the ASSERT_SERIES macros, which is included.