Centralize block creation statistic from Make_Series
As it costs some small amount to keep the stats up to date, it's
probably something that should be a debug-build only feature.  But
in any case, the statistic is now best kept in Make_Series when it is
told to use MKS_BLOCK...as opposed to repeating the line in other
routines that use the flag.