Avoid using Data_Stack for temporary REBVAL storage
As pointed out by @HostilFork, saving out form Do_* in an expandable
series is unsafe: the pointer could become dangling if it expands.
Because we don't [UN]SAVE_VALUE, use a length of 1 series for this
purpose. MKS_LOCK is not necessary here, as it's clear that this series
is not expanded, but in general, it should be used for series used for
this purpose.

The Data_Stack usage in callback_dispatcher should be fine. But, for
consistency, it's also changed.