ANY-LIST! typeset (and ANY-ARRAY! experiment)
Something that's been running a long time in the discussion is the
confusion resulting from "Block" being used interchangeably for
what an "ANY-BLOCK!" represents, as well as a term specifically
for the square-bracketed type of block. It is very confusing for
beginners, and for anyone trying to make a stab into the C source
code it is a barrier when reading almost anything. For instance:

    Val_Init_Paren(value, Make_Block(...))

One solution would be to put lots of "ANY" on the cases where you
mean to suggest you are making an Any_Block. That doesn't really
dig you out of the problem--it's still confusing--but now the names
are just a bit longer and you can technically argue it's more correct.

This goes ahead and moves on the choice that has been around
for a while now with no better options having come up: LIST!.
List has some data structure connotations with the word "linked"
but it is used in many more general senses, most notably in Lisp
for this exact purpose. That lists can have nested lists where the
nested item can be any type (including list) is now accepted in
HTML and elsewhere. It's a definition that must be explained to
users, but unlike ANY-BLOCK! it is a cost that can be paid up
front--and not be paying for it with confusion in every sentence
spoken after that understanding is established.

A more spontaneous terminology choice is on the distinction of
those series which are *not* lists and cannot hold arbitrary values
but only hold all the same type (codepoints for strings, bytes for
binaries, etc.) Outside of its usage in the function that pre-builds
a series, "array" is available. It's not a completely alien way
to talk about array as the "simple" sequential structure where
every element has to be the same type. (VECTOR! in C++, but
that's a more intimidating word and it's taken anyway. VECTOR!
would be an ARRAY! under this definition, BTW.)

If this turns out to be bad for some reason it can be changed.
There is just a growing need for clarity and distinction in the C
code making it necessary to take action on this terminology
problem.  There's a priority in unstacking the odds set against
people being able to read and modify the C source.

ANY-BLOCK! is kept around and working for the time being.