Formal Copy/Clone semantics, some Block=>Array
This is a phased commit of work that was done as part of the formal
managed/manual memory model. As a result, it has some terms on
the routines like "_Managed" that don't apply yet (but will). The
motivation for the cleanup at this time is thus not obvious...but the
reason is that tactical references to MANAGE_SERIES() are needing
to go in the code, and this helps see where they are needed.

The previous "Copy" and "Clone" routines were opaque in terms
of cues of what was actually happening. This breaks down each
copying primitive applied to arbitrary values into the semantics of
what it is actually doing.

    REBSER *Copy_Block_Len(block, index, len)

For instance, consider if you see `Copy_Block_Len(foo, 2, 4)`. You
have to figure out that 2 is the index into data, 4 is the length, and
read closely to see that the result may not be of length 4. Moreover,
foo may be a PATH! and its result could be going into a PAREN!,
with no BLOCK! in sight. Also, it is shallow copying when you might
have thought otherwise. The difference between copying and
cloning is not explained, and used somewhat randomly.

In this commit, the above operation is renamed:

    REBSER *Copy_Array_At_Max_Shallow(list, index, max)

Here the _At indicates the first parameter is the index at which
copying should start. Using _Max instead of _Len indicates that
the length may be longer than the available content, and you could
get a shorter result than what you pass in. (The name part "_Len"
indicates a contract that the elements will be available.) _Shallow
helps the author and those reading the callsite know that only a
shallow copy will be made. _Array instead of _Block helps with
the understanding that the series may actually be for a PATH!
or PAREN! type.

So while the longer names are slightly annoying, they reduce the
need for comments and can be macroized to drop the name part in
specializations. (e.g. dropping _At in the macro name and making
the index always 0).

This is a "portrait of the reforms", which will continue to be pushed
outward. But to keep the commit from getting too large to review,
this sticks to only Copy routines.