Add a global END_VALUE
Sometimes there is a moment where you need to pass a
pointer to a list of values that is terminated by a REB_END value
and you don't have any array to pass.  Using NULL as a no-op
requires the routine to be prepared to allow the null vs. just
looking for a REB_END, which it may not be.  This makes a
global end value you can use in such situations.

Because end is treated specially, it can't go where the other
such values live in ROOT...because END signals the end of the
root block.  It would be *technically* possible to use that block's
END value's address, as long as it didn't resize.  But that's a
bit sketchy.