Debug check and compliance for valid arg indexes
The bug in 7cb1104 drew attention to the fact that address sanitizer
isn't (always) catching accesses to out-of-bounds function arg
addresses, so this adds a debug check for it.  It's a little bit more
aggressive because it won't give you a pointer to an out of range
address and trust you not to derference it...it alerts when you ask.
Thus it required adjustment to existing code that asked for
parameters it didn't really need.

This commit also kills a no-longer-useful stack stressing assert.