Single-element paths are their element's evaluation
The implementation did not check for the case of a single element
path.   So when it tried to raise an error, the error object wrongly
tried to use an end of series terminator (a REB_END) in the
message.  This gave an assert.

Rather than turning the assert into a proper error, this elects to
make the path evaluate to the evaluation of the first element.  So:

    >> obj: context [t: now] get to-path 'obj
    == make object! [
        t: 26-Aug-2015/5:05:10-4:00
    ]