Make_Block() => Make_Array, => Is_Array_Series()
This continues the Block => Array change, this time with a more
basic renaming (isolated into a separate commit so as not to
make it harder to vet changes in the one about copy routines.)

As long as it's being renamed anyway, the IS_BLOCK_SERIES()
macro becomes another guinea pig for a second transformation,
whereby macros that act equivalently to inline functions w.r.t.
argument evaluation do not have to use all-caps casing.