Assert_"Blk" macro => Assert_Array
In the pre-Coherence-One era, when facing the problem of how
to name something that took an ANY-BLOCK! I had a few
routines that took a cue from Rebol's BLK_HEAD style macros
and would call these "BLK" routines intead of "Block" routines.

That's lame--and good to get rid of one of the holdovers of that
which I was responsible for repeating.