Differentiate GC-managed series from unmanaged
(Note: this commit adds a noisy alert which is sensitive and will
trigger if it sees anything suspicious.  There will likely be a period
of time when these alerts are discovered in code that hasn't been
run under the new rules.  Problems will be easiest to triage if you
are running Address Sanitizer, because they will pinpoint the
source of the problem...so ask how to build w/it if you're not yet.)

Previously all series were visible to the garbage collector from time
of allocation, yet there was still a Free_Series routine that could be
used.  That could lead to conditions where you might sometimes
successfully free a series before the GC got to it...and sometimes
it might beat you to the punch and crash.  It also meant it was extra
work to create a temporary series that wouldn't be considered by
the GC.

This is a simple and low-cost method of tightening the condition on
who is responsible for freeing a series memory.  Make_Series will
create a manually-memory managed series that must either be
freed with Free_Series or delegated to the garbage collector with
MANAGE_SERIES.  In a release build all this does is sets a flag
on the series...but in a debug build the series is also removed
from a tracking list.  This tracking list makes it easy to find out
when a function has "leaked" a series (in the sense of neither
freeing it or delegating it by the time it finishes running.)  There is
also protection in the sense that Free_Series will refuse to free any
series that has been delegated to the garbage collector.

To keep from having to add MANAGE_SERIES() calls to nearly
every place a series is allocated in the codebase, a trick is used.
The trick is to assume the moment of putting a series into a REBVAL
signifies that it should be the responsibility of the garbage collector.
That means the only places that really have to worry about it are
those that use temporary series which are not handed back to user
code.  (CALL used several of these series, and so needed the most
modification to deal with the change.)

This eliminates the GC_Infants array, which was a sketchy way of
approaching the desire by protecting a small fixed number of
recently allocated series from being freed.  (Which basically is just
a haphazard way of making bugs harder to find.)