Fix uninitialized Do_Path call in u-dialect.c
The interface for Do_Path was changed to take a pointer to a const
pointer for the path value, which is updated to reflect the path
position at termination.  (Under the current implementation, might
not be the end of path...because finding a function means giving back
the path position of the first refinement.  This is pending change.)

In any case, previously it had taken a pointer to a mutable pointer.
While you can pass a mutable pointer to a slot taking a const pointer
of the same type, you cannot pass a pointer to a mutable pointer to
something taking a pointer to a const pointer.  :-/

This was a case where an attempt to proxy the mutable value pointer
by way of a temporary const pointer failed to actually initialize the
proxy...so it was uninitialized.  DELECT isn't in the test suite or
commonly used so it wasn't noticed.

Found by Coverity.