Fix OS_STR* casting from/to host-specific REBCHR*
REBCHR is intended as host-specific character type. OS_STRCHR goes to
lengths to ensure that the REBCHR* argument is casted to the appropriate
platform-specific character types, but doesn't do the same for the
returned value. This, rightly, makes G++ 4.6.3 unhappy in release
builds:

    ../src/core/n-io.c: In function 'REBSER*
    String_List_To_Block(REBCHR*)':
    ../src/core/n-io.c:730:53: error: invalid conversion from 'const
    char*' to 'REBCHR* {aka unsigned char*}' [-fpermissive]

Second problem is the C++-specific overload of strchr which returns a
const pointer when passed in a const. This was already tackled in the
debug OS_STRCHR by casting away the const, we now do the same for the
release OS_STRCHR.

The same const overload also exists for wcschr, so we add a similar
mutability cast for the wcschr-using OS_STRCHR variant on OS_WIDE_CHAR
platforms.

Finally, the from/to REBCHR casting was completely missing from the
other OS_STR* defines on OS_WIDE_CHAR platforms.