Remove too-aggressive leak check asserts for trap
When you PUSH_TRAP it sets up a state. In that state a debug
build keeps track of the list of manually memory-managed series
that are in play.  This list is checked when DROP_TRAP runs to
make sure there was no net accumulation between the PUSH
and the DROP--and that every new series created was either
handed to the GC to manage or freed manually.

There were two additional places where this assert was made:
at the point of throwing an error and catching it.  It would make
the assert using the list of manuals as captured in the state at
the time of the push.  This is too aggressive, for instance:

     PUSH_TRAP()
     series = Make_Series(...);
     if (error) { handle trap, Free_Series(series), return }
     Code_That_Might_Trap(series);
     Free_Series(series);
     DROP_TRAP();

That should be legal.  But with the asserts written as they were,
it would have been seen as a leak at the time of trapping,
because series was not GC managed and hadn't been freed
yet.  (But it was going to be!)

Checking at DROP_TRAP() is good enough, so this deletes the
checks in Do_Error() and Trapped_Helper_Halted().