NORETURN attribute macro, DEAD_END => __builtin
This adds a section for attribute macros in reb-c.h, and adds the
NORETURN attribute.  It also changes DEAD_END to use the
__builtin_unreachable function in GCC and Clang.

This is a prelude to the new philosophy of no longer using fake
returns to second-guess the compiler's "not all paths return a
value" logic, or what the analysis tools like Coverity find.  It was a
losing battle because you can't write a macro that works for all
situations... return 0 won't work in void functions, and it also won't
work for things that return structs by value.  So we'll only work to
appease the compiler warnings on those compilers that can
respect some notion of NORETURN and DEAD_END--otherwise
just turn the warnings off.