Avoid leak in case of creating error with invalid #
When creating an error with an invalid ID number (such as 0), then
Make_Error() will fail...but it already has created a partial series for
the error it was trying to create.  This frees that partial before raising
the creation error so it doesn't leak.  (Would formerly have been
picked up by the GC, but it's not managed yet...)