Flip convention on stray Do routines to TRUE if throws
Switching to returning a boolean from routines that could throw
while executing Rebol code had a number of advantages.  Initially
that was to prevent the need to do two checks for if something was
a REB_ERROR and if it was THROWN().  Now it removes just one
check on the THROWN() flag...but it means that the thrown flag
could potentially be taken off the value altogether.

Another value is to emphasize at the callsite of the routine that
there is a result to process.  Initially, that result was biased so that
TRUE meant that the call was successful, so `if (Do_Thing())...`.
However, the special handling was needed not in the case of success,
but whenever the result was thrown.  With Do_Block_Throws the
convention was established that TRUE would be used in the throw
case and you could read is `if (Do_Block_Throws(...))` which gave a
readable flow to the code.

This picks up some routines that were lingering on the TRUE for
success convention, and also makes sure they are all inside of an IF
that does *something* with the thrown case.