foreach => for-each/every, foreach as legacy
FOREACH was a bit of a black sheep as not using a hyphen
to separate the words, and represented a kind of slippery slope
that had been rejected by Rebol's philosophy (TYPE-OF and not
TYPEOF, for instance).  It also was in a family of other functions
that did have a hyphen, like remove-each and map-each.  It
also draws the eye toward the componet REACH..which once
it is seen sort of can't be unseen.

Debate centered on whether it was common enough to warrant
a non-hyphenated version, and just call it EACH.  That did not
seem to sit well, but the non-hyphenated synonym EVERY came
out as a possibility.  As there are other synonyms, it seems worth
a try.  So speculatively this adds that.

It also keeps FOREACH in the legacy mezzanines.