Inline Do_Args into Do_Core
There are a number of benefits to this change, some related to
optimization and reducing the stack depth.  But mostly that when
you have a piece of functionality that runs exactly one place, the
inlining of it lets you see what's going on better.  So the loop
gets easier to optimize, and also to perhaps break up in a more
useful way.

(Although: a long function is not in itself a bad thing, because if a
bit of functionality is called exactly once in a program then breaking
it into a function actually makes the program *longer*...)

Expanded comments also.