Make an ASSERT_SERIES_MANAGED macro
There were some places in the code that just assert()ed on
the managed status, which is not as informative as having a
Panic_Series.  Some cases added the Panic_Series manually,
and had a bit of a descriptive string...but the line number and
file are good enough.  So switched them to use the macro too.