Properly handle errors in callback!
Callback! calls Do* to execute the code, which might throw an error by
calling longjmp. However, FFI is not able to handle this, but it's OK if
the setjmp and longjmp happen in a single FFI call (
http://www.freelists.org/post/luajit/FFI-and-setjmplongjmp,1):

    "It's certainly not safe if the setjmp happens in one FFI call and
    the longjmp happens in another one. If everything happens inside a
    single FFI call, then yes, that would be ok."

A global variable is introduced to save the potential error from the
callback!. This needs to be global, because callback_dispatcher only
takes on pointer as the user_data, and this user_data is pointing to the
callback, when callback_dispatcher is called in Call_Routine, where the
error is raised to upper level, there is no access to the callback!,
thus the error! and callback! can't be connected, and a global variabl is
required.