Allow Do_Core on an unmanaged block
There are assertions in place to make sure that all values
that are exposed to users are under control of the GC.  This
is a good thing.  But it also asserted that the series being
given to Do_Core to execute be managed by the GC.

This is too aggressive, as there are useful scenarios in which
a temporary series is created...executed...and freed.  There
are a few of these in Ren/C, and if they were forced to use
managed series then operations would be accumulating
series even when they don't need to.

Other usages for the feature will likely arise.  It's okay so long
as the series is private, because Do_Core on a series does
not give that series out to the user...they have to get ahold of
it some other way.  If you can assure there is no other way,
executing unmanaged series should be allowed.