Clarify GC managed vs. not in UTF8 conversions
When a UTF8 series was needed from a string, Rebol was
aggressive in not making a new series if the original string
was ASCII (as if all the bytes in a string are < 128, then it is
legal UTF8 already).  This resulted in an interface that would
sometimes allocate a new series and sometimes not.

That complicates making it clear in reading the code as to
whether a series is not yet visible to GC (and must be freed) vs.
when it is visible to the GC (and must be protected across any
calls to the evaluator or I/O if it shouldn't be garbage collected).

This pins down the interfaces. Make_UTF8_From_Any_String()
always creates a new series, and Temp_Bin_Str_Managed()
returns a managed series which may-or-may-not be the series
you passed in (which is asserted to be managed).  The former
case thus always requires freeing, while the latter never does.

Also this adds comments to explain better what Temp_Bin_Str
is for, and why its variance in behavior makes it likely not
something you want to put into a value and hand back into user
code.  (It had been called "Prep_Bin_Str" previously.)