Make SERIES_LEN() a synonym for SERIES_TAIL()
In C programming, the general understanding of when you ask
for the length of something--even a null terminated string--is that
the "LEN" (strlen(), whatever) does not include any terminators,
but is the count of the elements.

That's generally true in Rebol too where _LEN arises, except for
some reason in SERIES.  To get the length is SERIES_TAIL if you
don't want the terminator, which is a bit confusing because that
seems it would refer to a position and not an index.

There were only a few references to SERIES_LEN but a *lot* of
references to SERIES_TAIL (with a mixture of whether the
variables related were called `len` or `tail`).  Rather than do a
replace right now, this makes SERIES_LEN a synonym for
SERIES_TAIL to eliminate confusing uses.  Eventually all the
tail references should be switched to LEN.