Don't zero fill memory allocations by default
Rebol's wrapper for memory allocation internally is called
Make_Mem.  I've gone in and commented about why it's done
the way it is (or why I think it is, anyway).  One thing it would do
was that it would zero memory by default.

This takes that off, which brings two benefits.  One benefit is
not paying for the initialization if those zeros are just going to be
overwritten anyway.  The other more important benefit is that
it means address sanitizer and valgrind can tell you when you
are accessing uninitialized parts of an allocation...because once
it sees you've written a zero it assumes the memory is "valid".

One area where memory still has to be cleared in the current
implementation is a "Node" allocation from a series pool, when
it first creates a reusable node (but not on each reused allocation).
Added comments to explain what's going on and what directions
might be taken with it, but for now leaving it to work how it does.