Eliminate Guard_Series and Loose_Series
Because it's necessary to protect series from garbage collection
when you know they aren't reachable from the root set, there
is Save_Series and Unsave_Series.  These routines are for
temporarily keeping a series alive, and use an efficient stack
approach for pushing and popping the protections.  If an error
is trapped in the middle of a state, then after the longjmp any
protections in the stack will be released.

Guard_Series and Loose_Series were similar, except without
a stack requirement...so you could just guard any series for an
unlimited amount of time.  This made the removals in
Loose_Series less efficient, because it would have to search
a list to find the series to release.  It also meant that there was
now a way to mark a series as being live persistently with no
reference from the root set...and leak.

In fact the one call to Guard_Series was from port code that
would wind up leaking a REBREQ-sized series on every
allocation...because it never called Loose_Series.  It does not
appear to have been necessary because the REBREQ it is
storing refers to the port...so as long as it is reachable by the
set already covered by the GC (which it should be).

It may be necessary to have some more ways to protect things
from the GC, but a carte-blanche like Guard_Series to request
a "forever" protection with no checking isn't a good shape to
encourage.  So I'm removing it as it seems that it did not create
problems for the REBREQs, but we will have to see if that is the
case for everyone's code.