Fully manage object in reflecting datatype specs
The managed/unmanaged distinction for series helps to tell
you which way to worry...either worry about a series being
GC'd out from under you, or worry about leaking.

Objects present a special case because they are made up of
two series: a "frame" and "words".  When making copies of
an object, you may well want to keep the words list when you
copy (so long as you're not adding key/value pairs to the
object, in which case you can't use the same frame).  To help
keep track of things, Val_Init_Object notifies you if the
managed state of your words and frame don't match.

The value of this is questionable, but for now it means that
anyone who deals with an object that isn't either fully managed
or fully unmanaged will have their attention drawn to it.
This fixes a case where DATATYPE! reflection didn't line
up the management state before Val_Init_Object (seen with
an assertion if you typed `spec-of integer!`)