Fully enforce the "no exit or quit during Init" rule
In trying to tighten up a "clean shutdown" for the interpreter on
all code paths, one place that was willing to shutdown in what
would be called an "incomplete state" is during Init_Core.  The
bit of code that runs at the end of startup could mechanically
QUIT or EXIT, but the Init_Core routine returns void and has
no way to return the exit status.

In the general case, it may be desirable that if Init cannot finish
there be a way to throw a safe exception to the host.  However,
those failures should be due to insufficient memory or other
conditions, raised as errors...and not THROW-style QUIT or
EXIT requests.  This enforces the rule and narrows the number
of what should be checked as "clean exits" from the program.