Add simple Shutdown() complements for Init()s
These are some easy-to-understand routines to use during
shutdown that don't really do anything other than free items that
are consuming memory.  They are just complements to the Init()
functions, and are placed in the files next to those functions.

It also comments in Shutdown_Core on why this accounting is
important, even though often in a release build a program can just
exit to the OS and not worry about freeing each allocation.