Add "shutdown" flag to GC, call it from Shutdown
This adds a shutdown flag to both Sweep_Series and Recycle
(and does a little bit of code cleanup and commenting too.)

Preliminarily, this makes sure the GC runs properly when
there is nothing in the root set...and then does a manual
sweep to free any non-GC managed series.  This operates
under the assumption that when there is no root set, all non
series things will be collected...with only manual series left.