Free the memory pools during Shutdown
This shuts down the memory allocation system itself, and thus
needs to be one of the last shutdown routines.  Because it has
to be last, the Prior_Expand list allocation and deallocation is
moved into m_pools.c as well, because that cannot run until
after all series are freed...so it's a good time to do it.  (There
may still be manual series outstanding after the GC is shut down,
so it had to be moved out of m-gc.c)

There is an assert which is currently commented out.  That is
regarding the number representing Rebol's accounting of how
much memory it has...which should balance to 0.  It doesn't yet,
but hopefully will in the coming commits.