Replace RL_Reset with RL_Shutdown
There was a vaguely named API called "RL_Reset" in RL_API,
which said it would "reset the REBOL interpreter".  Given that it
was unimplemented, it is probably replaced by being able to do
a shutdown of the interpreter and then calling init again.

The newly-added RL_Shutdown routine adds a speculative
flag to say whether it is a "clean" shutdown or not.  If it is not a
clean shutdown then the assumption is that the caller does not
need the memory and such reclaimed, and a faster shutdown
process that does only the essentials can be run.  As it is, there
are no "essentials" so really that just means not calling shutdown,
but there may be some at some point.  Debug builds do a clean
shutdown regardless.

This adds a call to the shutdown which is (currently) never
reached, because the RL_Do_String function call will quit to the
OS without offering the host the chance to do their own shutdown.
This will be addressed in further commits.