Free leaked `home_dir` set by Parse_Args
The Parse_Args function uses the OS_Get_Current_Dir call
to set the home_dir field in the Main_Args structure, and this is
detected as leaked on program exit.  Generally speaking the
responsibility for these structures and their memory needs to be
thought out in a more systemic way, but this addresses the
problem for the moment and draws attention to the need for a
better answer.