Give RL_Print_TOS mold option, add RL_Drop_TOS
This isn't an attempt to redesign the RL_API to make it "good",
it just happened to be that a need for the RL_Drop_TOS API
came up again.  Since the need already existed, this breaks
it out of RL_Print_TOS so you can drop without printing (or
print without dropping).