Shift console control to host-main from RL_Api
This is an overdue push to try and do some amount of
reform of the RL_Api, enough to the point that the host-main.c
file can show a reasonable control flow from start to finish of
the utility that people run with "rebol.exe".

It may be that REBRXTs can be abandoned entirely, but for
the moment this tries to keep them while changing the API
slightly to give an exit code (needed to be returned by the
console).  It also formalizes the differentiation between errors,
halting, exiting, and getting back a normal value from the
evaluator.  This puts the host in a bit more control about the
presentation for those different results (though printing is still
deferred to "wherever Rebol is putting its output").