BREAK/CONTINUE in condition of while rethrows
As written currently, a WHILE loop will process breaks and
continues that happen in its condition.  One might disagree with
how it handles them even so:

    x: 0
    while [either x = 0 [x: 1 continue] [false]] [
        print "Hello"
    ]

That will print Hello, indicating that the CONTINUE did not jump
to processing the beginning of the condition again.  Instead it
jumped to the body.

But arguing over that is probably less relevant than observing
that the condition should probably be considered to not be a
candidate for traps of BREAK and CONTINUE by the while,
rather to any enclosing catchers of BREAK and CONTINUE.
This is what Red does now (either intentionally or unintentionally)
and seems the better option.

One thing that makes it better is that it gives an opportunity to
looping code that wants to break out of the enclosing context
to have a chance to do so in the condition.  Hence it is the
"choice of greatest expressivity".

Also cleans up code and makes unset handling clearer.