Switch C stack overflow "check" to just the test
The function CHECK_C_STACK_OVERFLOW doesn't quite
convey the consequence of what will happen (namely raising an
error), and it also doesn't give options...such as doing cleanup
before raising the error (or handling it somehow and not raising
an error at all).

This switches it to being just a test, with the response at the
callsite (which is generally Trap_Stack_Overflow()).