Free extant manual series during Trap()
This was a question early on, as to whether it should be the
responsibility of code using series to find a way to free the
manual series they'd allocated if an error was going to be
trapped... or if it should be handled automatically.  Handling it
automatically meant that some extra tracking would have to be
in the release build--while not handling it meant the tracking
would only be in the debug build for alerts.

The rationale for not doing it automatically was that in C code,
you cannot do a longjmp in the middle of arbitrary code and not
expect to leak resources.  There is no stack unwinding and no
destructors...so if an error traps while you've got a file open,
then you have a leaked file handle on your hands.  Given that
no general mechanism was in existence, it seemed to suggest
vigilance on the part of coders to make sure they didn't do a
Trap unless they were sure it was safe.  That made it cheaper
to mark a series managed, since there was no tracking list.

However, the nail in the coffin for this was out-of-memory errors
and C stack overflows.  These can occur in routines and at times
over which there isn't as much control...and getting the control
would involve putting in checks at every allocation or copy.  So
automatic handling would have to be added.

In order to make the cost acceptable, the linked-list approach
(which cost two extra pointers per series in the debug build)
had to be replaced.  The new code uses a list of series pointers
representing the extant manuals.  This list is pushed at the back,
and when removals are done it searches from back to front.
Most of the time the manual series is going to be the last one
allocated, but if it isn't then it will go find it and then swap the
positions of whatever is last into that slot and drop the tail.

Doing automatic freeing of extant manual series will not be
able to magically solve all the exception handling needs.  If a
resource that needs to be cleaned up isn't a Rebol series,
one will need to do a PUSH_TRAP and have handling code.
That way you can respond to out of memory errors or a stack
overflow resulting from things like copying recursive blocks.
But this at least takes care of the common need and makes
it easier whether you push a trap or not.