Remove MANAGE_SERIES from cases just for trap
There were a couple of places in the FFI code that were
doing a MANAGE_SERIES solely for the reason of protecting
the series from a trap, when it was an internal series that
could be freed without keeping it around.  With the change
that unmanaged series will still be cleaned up in the case of
a trap, they could be changed to not manage and just use
an ordinary free.

One case could not be changed due to the current
requirement that DO has, that any block it is executing must
be managed.  I actually confronted this requirement and may
need to ease it for the implementation of Ren/C++... so I went
ahead and structured the code so that it could be changed to
not MANAGE_SERIES() if that happens (and commented it
to that effect).