Add Init/Shutdown for the host extensions
The "host extension" mechanism is probably not something
that has a place in the picture long term... but it is there for now
and this adds a place where the console can call it to clean up
allocations it has made.

It also removes the dead "licensing" host extension reference
and fixes a leak of calling the RL_API for getting the words of
an object.