Fix block cipher leak in TLS protocol
The stream ciphers used in TLS (RC4, AES) required a malloc'd
unit of state.  It would be allocated on the first call, and then you
would pass it in each time you had some more data to get or
put over the line.

Part of the protocol was that when you were done, you would
pass in a NONE! so that the state could be freed.  But that
call was never happening.  This adds it into the TLS actor's close
method, which seems to take care of the leak.