Simplify shared routine for BREAK/CONTINUE
The loop constructs had a confusing shared routine called
Process_Loop_Throw() that returned an integer result, which
has been steadily getting less confusing.  This moves the
shared routine out of c-error.c (as BREAK and CONTINUE
are throw-class items which are no longer ERROR!s) and
makes it return a simple answer of whether a received throw
suggests needing to return out of the loop processing function
or not.

One case that had to be looked at specially was MAP-EACH,
which is sensitive to whether you call it with BREAK or a
BREAK/WITH.  In the case of BREAK it is assumed to mean
that you still want to return the accumulated result (vs. the
UNSET! that a BREAK returns by default).  But if you
BREAK/WITH then you are overriding the return value from
that accumulated result with a new value.

(Note: Because the thrown break itself doesn't remember
whether you used /WITH if the value is an UNSET!, this means
that BREAK/WITH a literal UNSET! will not return unset from
the map-each...it will return the thus-far accumulated result.)